organise charcoal pit classes
