+++ src/main/java/net/jmb19905/CarbonizeClient.java
@@ -2,7 +2,7 @@ package net.jmb19905;
 
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.model.loading.v1.ModelLoadingPlugin;
-import net.jmb19905.model.CharringWoodBlockModel;
+import net.jmb19905.charcoal_pit.block.CharringWoodBlockModel;
 import static net.jmb19905.Carbonize.*;
 
 public class CarbonizeClient implements ClientModInitializer {
